+++ demo/admin/src/products/future/ManufacturerFilter.tsx
@@ -1,8 +1,6 @@
 import { gql, useQuery } from "@apollo/client";
 import Autocomplete from "@mui/material/Autocomplete";
-import TextField from "@mui/material/TextField";
 import { GridFilterInputValueProps, GridFilterOperator } from "@mui/x-data-grid-pro";
-import { GQLManufacturersFilterQuery, GQLManufacturersFilterQueryVariables } from "@src/products/future/ManufacturerFilter.generated";
 import * as React from "react";
 import { useIntl } from "react-intl";
 
@@ -12,6 +10,11 @@ const manufacturerFilterFragment = gql`
         name
     }
 `;
+
+import { InputBase } from "@mui/material";
+import { useDebounce } from "use-debounce";
+
+import { GQLManufacturersFilterQuery, GQLManufacturersFilterQueryVariables } from "./ManufacturerFilter.generated";
 
 const manufacturersQuery = gql`
     query ManufacturersFilter($offset: Int!, $limit: Int!, $search: String) {
@@ -26,14 +29,16 @@ const manufacturersQuery = gql`
 `;
 
 // Source: https://mui.com/x/react-data-grid/filtering/customization/#multiple-values-operator
-function ManufacturerFilter(props: GridFilterInputValueProps) {
-    const { item, applyValue } = props;
+function ManufacturerFilter({ item, applyValue }: GridFilterInputValueProps) {
     const intl = useIntl();
+    const [search, setSearch] = React.useState<string | undefined>(undefined);
+    const [debouncedSearch] = useDebounce(search, 500);
 
     const { data } = useQuery<GQLManufacturersFilterQuery, GQLManufacturersFilterQueryVariables>(manufacturersQuery, {
         variables: {
             offset: 0,
             limit: 10,
+            search: debouncedSearch,
         },
     });
 
@@ -44,7 +49,8 @@ function ManufacturerFilter(props: GridFilterInputValueProps) {
             size="small"
             options={data?.manufacturers.nodes ?? []}
             autoHighlight
-            value={item.value}
+            value={item.value ? item.value : null}
+            filterOptions={(x) => x} // disable local filtering
             isOptionEqualToValue={(option, value) => {
                 // does only highlight the selected value in options-list but does not trigger getOptionLabel-Call
                 return option.id == value;
@@ -55,15 +61,17 @@ function ManufacturerFilter(props: GridFilterInputValueProps) {
             }}
             onChange={(event, value, reason) => {
                 // value can't be "{ id: value.id, name: value.name }" because value is sent to api
-                applyValue({ id: item.id, operatorValue: "equals", value: value ? value.id : undefined, columnField: "manufacturer" }); // columnField needs to match field-prop and filter-name
+                applyValue({ id: item.id, operatorValue: "equals", value: value ? value.id : undefined, columnField: "manufacturer" });
             }}
             renderInput={(params) => (
-                <TextField
+                <InputBase
                     {...params}
+                    {...params.InputProps}
+                    autoComplete="off"
                     placeholder={intl.formatMessage({ id: "manufacturer-filter.placeholder", defaultMessage: "Choose a manufacturer" })}
-                    inputProps={{
-                        ...params.inputProps,
-                        autoComplete: "new-password", // disable autocomplete and autofill
+                    value={search ? search : null}
+                    onChange={(event) => {
+                        setSearch(event.target.value);
                     }}
                 />
             )}