+++ packages/admin/admin/src/dataGrid/excelExport/generateExcelFile.ts
@@ -46,21 +46,22 @@ export function generateExcelFile<Row extends GridValidRowModel>(
                     let value = row[column.field];
                     if (column.valueGetter) {
                         // @ts-expect-error `valueGetter` requires more data but we don't have all that data available so we only pass in what we have and hope nothing breaks
-                        value = column.valueGetter({ value, row });
+                        value = column.valueGetter({ value, row }) ?? "";
                     }
                     if (column.valueFormatter) {
                         // @ts-expect-error `valueFormatter` requires more data but we don't have all that data available so we only pass in what we have and hope nothing breaks
-                        value = column.valueFormatter({ value });
+                        value = column.valueFormatter({ value }) ?? "";
                     }
-                    if (typeof value !== "string" && typeof value !== "number" && value !== null) {
-                        throw new Error(`Provided value must be of type string, number or null but is "${typeof value}"`);
+
+                    if (typeof value !== "string" && typeof value !== "number" && value !== null && !(value instanceof Date)) {
+                        throw new Error(`The type of the provided value "${typeof value}" is not supported for excel export.`);
                     }
 
                     excelRow[column.field + columnIndex] = value;
                 }
             });
 
-            worksheet.addRow({ ...excelRow });
+            worksheet.addRow(excelRow);
         } catch (e) {
             throw new Error(e);
         }