+++ storybook/src/docs/components/Table/07_TableDndOrder.stories.tsx
@@ -1,6 +1,5 @@
 import { TableDndOrder, TableLocalChanges } from "@comet/admin";
 import { Button } from "@mui/material";
-import * as React from "react";
 
 import { dndProviderDecorator } from "../../../dnd.decorator";