+++ storybook/src/docs/helper/clipboard.stories.tsx
@@ -1,6 +1,5 @@
 import { readClipboardText, writeClipboardText } from "@comet/admin";
 import { Button } from "@mui/material";
-import * as React from "react";
 
 export default {
     title: "Docs/Helper/Clipboard",